henter fra database og tester
