+++ app/src/androidTest/java/DatabaseHelperTest.java
@@ -1,0 +1,2 @@
+public class DatabaseHelperTest {
+}
