+++ app/src/androidTest/java/com/elinstallator_hold4/DataBaseHelpertest.java
@@ -1,0 +1,9 @@
+package com.elinstallator_hold4;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import junit.framework.TestCase;
+import org.junit.runner.RunWith;
+
+@RunWith(AndroidJUnit4.class)
+class DataBaseHelpertest extends TestCase {
+}
